fix: properly support mocha equire API
Fixes https://github.com/microsoft/vscode-test/issues/259
